+++ packages/eas-cli/src/commands/worker/alias.ts
@@ -35,9 +35,6 @@ interface RawDeployAliasFlags {
 export default class WorkerAlias extends EasCommand {
   static override description = 'Assign deployment aliases';
   static override aliases = ['deploy:alias', 'deploy:promote'];
-
-  // TODO(@kitten): Keep command hidden until worker deployments are live
-  static override hidden = true;
   static override state = 'beta';
 
   static override flags = {